added base estimator to is fitted state and checks
