+++ sktime/base/__init__.py
@@ -3,4 +3,5 @@
 
 __author__ = ["Markus Löning"]
 
+from sktime.base._base import BaseEstimator
 from sktime.base._meta import BaseComposition
