+++ sktime/forecasting/base/_base.py
@@ -8,7 +8,7 @@ __all__ = [
     "is_forecaster"
 ]
 
-from sklearn.base import BaseEstimator
+from sktime.base import BaseEstimator
 
 DEFAULT_ALPHA = 0.05
