+++ sktime/forecasting/tests/__init__.py
@@ -16,4 +16,4 @@ DEFAULT_INSAMPLE_FHS = [
     np.array([-3, 2])  # mixed in-sample and out-of-sample
 ]
 DEFAULT_SPS = [3, 7, 12]
-DEFAULT_ALPHAS = [0.95, 0.9]
+DEFAULT_ALPHAS = [0.99, 0.95]
